+++ src/assets/Icons/index.ts
@@ -1,13 +1,15 @@
-import bagIcon from './bag.svg';
-import searchIcon from './search.svg';
-import userIcon from './user.svg';
-import logoIcon from './logo.svg';
-import more from './more.svg'
+import about from './about-logo.svg'
+import appStore from './appStore.svg'
+import bagIcon from './bag.svg'
+import googlePlay from './googlePlay.svg'
 import ils from './ils.svg'
 import lineIcon from './Line.svg'
+import logoIcon from './logo.svg'
+import more from './more.svg'
+import searchIcon from './search.svg'
 import spicy from './spicy.svg'
+import userIcon from './user.svg'
 import vegan from './vegan.svg'
-import vegetarian from './vegetarian.svg'
+import  vegetarian  from './vegetarian.svg'
 
-
-export {bagIcon,userIcon,searchIcon,logoIcon,more,ils,lineIcon, spicy, vegan, vegetarian }
+export {about,appStore,bagIcon,googlePlay,ils,lineIcon,logoIcon,more,searchIcon,spicy,userIcon,vegan,vegetarian,}
