+++ src/styles/About.scss
@@ -4,7 +4,7 @@
     gap: 30vw;
     width: 100vw;
     max-width: 100vw;
-    height: 80vh;
+    height: 60vh;
     background-color: #FAFAFA;
     align-items: center;
     justify-content: center;
@@ -30,11 +30,11 @@
     align-items: flex-start;
     gap: 2vw;
     img{
-        width: 13vw;
+        width: 10vw;
     }
 }
-h3{
-    font-family: Helvetica;
-    font-weight: 200;
-    font-size: 2vw;
+.about-logo{
+    img{
+        width: 10vw;
+    }
 }
